ok, ignore browsergym eval test on regular CI
